Add files via upload
K-means clustering is a method that comes from signal processing and uses the k-means algorithm. It aims to group a set of n observations into k clusters. Each observation is placed in the cluster where the average (or center) of the cluster is closest to it, making that cluster its representative.